WOLO-164-restricted button only, better error message
