+++ src/styles/sign-in-section.scss
@@ -1,0 +1,6 @@
+@import './common.scss';
+
+#sign_in_section_error{
+    color: $color-4-red;
+    margin: 10px;
+}
